+++ scripts/test/Selenium/Customer/CustomerPreferences.t
@@ -24,6 +24,15 @@ $Selenium->RunTest(
         # get helper object
         my $Helper = $Kernel::OM->Get('Kernel::System::UnitTest::Helper');
 
+        # enable google authenticator shared secret preference
+        my $SharedSecretConfig = $Kernel::OM->Get('Kernel::Config')->Get('CustomerPreferencesGroups')->{'GoogleAuthenticatorSecretKey'};
+        $SharedSecretConfig->{Active} = 1;
+        $Helper->ConfigSettingChange(
+            Valid => 1,
+            Key   => "CustomerPreferencesGroups###GoogleAuthenticatorSecretKey",
+            Value => $SharedSecretConfig,
+        );
+
         # create test customer user and login
         my $TestCustomerUserLogin = $Helper->TestCustomerUserCreate(
         ) || die "Did not get test customer user";
@@ -42,7 +51,7 @@ $Selenium->RunTest(
 
         # check CustomerPreferences screen
         for my $ID (
-            qw(UserLanguage UserShowTickets UserRefreshTime CurPw NewPw NewPw1)
+            qw(UserLanguage UserShowTickets UserRefreshTime CurPw NewPw NewPw1 UserGoogleAuthenticatorSecretKey)
             )
         {
             my $Element = $Selenium->find_element( "#$ID", 'css' );
@@ -123,6 +132,22 @@ $Selenium->RunTest(
                 "Test widget 'Ticket overview' found on screen"
             );
         }
+
+        # try updating the UserGoogleAuthenticatorSecret (which has a regex validation configured)
+        $Selenium->find_element( "#UserGoogleAuthenticatorSecretKey", 'css' )->send_keys('Invalid Key');
+        $Selenium->find_element( "#UserGoogleAuthenticatorSecretKey", 'css' )->VerifiedSubmit();
+        $Self->True(
+            index( $Selenium->get_page_source(), $SharedSecretConfig->{'ValidateRegexMessage'} ) > -1,
+            "Error message for invalid shared secret found on screen"
+        );
+
+        # now use a valid secret
+        $Selenium->find_element( "#UserGoogleAuthenticatorSecretKey", 'css' )->send_keys('ABCABCABCABCABC2');
+        $Selenium->find_element( "#UserGoogleAuthenticatorSecretKey", 'css' )->VerifiedSubmit();
+        $Self->True(
+            index( $Selenium->get_page_source(), 'Preferences updated successfully!' ) > -1,
+            "Success message found on screen"
+        );
 
         # Inject malicious code in user language variable.
         my $MaliciousCode = 'en\\\'});window.iShouldNotExist=true;Core.Config.AddConfig({a:\\\'';